refactor: 쿠폰 계산 로직들 커스텀 훅으로 분리
